refactor(ui5-bar): move to main
