--- packages/fiori/src/themes/base/sizes-parameters.css
@@ -1,6 +1,0 @@
-[data-ui5-compact-size],
-.ui5-content-density-compact,
-.sapUiSizeCompact {
-	--_ui5_bar_base_height: 2.5rem;
-	--_ui5_bar_subheader_height: 2.25rem;
-}
