+++ packages/fiori/src/themes/sap_belize/parameters-bundle.css
@@ -1,5 +1,3 @@
-@import "../base/sizes-parameters.css";
-@import "../base/Bar-parameters.css";
 @import "../base/FlexibleColumnLayout-parameters.css";
 @import "../base/MediaGallery-parameters.css";
 @import "../base/MediaGalleryItem-parameters.css";
