+++ packages/main/src/themes/sap_fiori_3_hcw/Bar-parameters.css
