+++ packages/website/docs/_samples/main/Bar/Basic/main.js
